Got tests working after restructuring
