+++ africanus/calibration/phase_only/__init__.py
@@ -1,4 +1,3 @@
 # flake8: noqa
 
-from africanus.calibration.utils import residual_vis, correct_vis
-from africanus.calibration.phase_only import phase_only_gauss_newton
+from africanus.calibration.phase_only.phase_only import phase_only_gauss_newton
