+++ africanus/calibration/utils/dask.py
@@ -6,12 +6,15 @@ from __future__ import print_function
 
 from functools import wraps
 
-from africanus.calibration.utils import CORRECT_VIS_DOCS, CORRUPT_VIS_DOCS
-from africanus.calibration.utils import RESIDUAL_VIS_DOCS
+from africanus.calibration.utils.correct_vis import CORRECT_VIS_DOCS 
+from africanus.calibration.utils.corrupt_vis import CORRUPT_VIS_DOCS
+from africanus.calibration.utils.residual_vis import RESIDUAL_VIS_DOCS
 from africanus.calibration.utils import correct_vis as np_correct_vis
 from africanus.calibration.utils import corrupt_vis as np_corrupt_vis
 from africanus.calibration.utils import residual_vis as np_residual_vis
 from africanus.calibration.utils import check_type
+
+from africanus.util.requirements import requires_optional
 
 try:
     from dask.array.core import blockwise
