+++ africanus/calibration/utils/tests/__init__.py
@@ -1,0 +1,1 @@
+# flake8: noqa
